Upgrade model

+++ src/handler.py
@@ -14,7 +14,7 @@ try:
     pipe = AutoPipelineForText2Image.from_pretrained(
         "SG161222/RealVisXL_V5.0_Lightning",
         torch_dtype=torch.float16,
-        variant="fp16",
+        variant="fp32",
         add_watermarker=False,
         use_safetensors=True)
     pipe.scheduler = DPMSolverSDEScheduler.from_config(pipe.scheduler.config)
